API

General functionalities

IteratorSampling.itsample — Function

itsample([rng], iter; kwargs...)

Return a random element of the iterator, optionally specifying a rng (which defaults to Random.default_rng()). If the iterator is empty, it returns nothing.

itsample([rng], iter, n::Int; replace = false, ordered = false, kwargs...)

Return a vector of n random elements of the iterator, optionally specifying a rng (which defaults to Random.default_rng()).

replace dictates whether sampling is performed with replacement. ordered dictates whether an ordered sample (also called a sequential sample, i.e. a sample where items appear in the same order as in iter).

If the iterator has less than n elements, in the case of sampling without replacement, it returns a vector of those elements.

Specific algorithms

IteratorSampling.reservoir_sample — Function

reservoir_sample(rng, iter; method = :alg_L)
reservoir_sample(rng, iter, n; replace = false, ordered = false, kwargs...)

Reservoir sampling algorithm with and without replacement.

The method keyword can be either :alg_L or :alg_R. The optional kwargs are passed to the more specific methods called internally by this function: reservoir_sample_without_replacement and reservoir_sample_with_replacement.

IteratorSampling.reservoir_sample_without_replacement — Function

reservoir_sample_without_replacement(rng, iter, n; ordered = false, method = :alg_L)

Reservoir sampling algorithm without replacement. The method keyword can be either :alg_L or :alg_R.

Adapted from algorithms R and L described in "Random sampling with a reservoir, Jeffrey S. Vitter, 1985".

IteratorSampling.reservoir_sample_with_replacement — Function

reservoir_sample_with_replacement(rng, iter, n; ordered = false)

Reservoir sampling algorithm with replacement.

Adapted fron algorithm RSWR_SKIP described in "Reservoir-based Random Sampling with Replacement from Data Stream, Byung-Hoon Park et al., 2008".

IteratorSampling.sortedindices_sample — Function

sortedindices_sample(rng, iter)
sortedindices_sample(rng, iter, n; replace = false, ordered = false)

Algorithm which generates sorted random indices used to retrieve the sample from the iterable. The number of elements in the iterable needs to be known before starting the sampling.