API

General functionalities

StreamSampling.ReservoirSample — Function

ReservoirSample([rng], T, method = algL)
ReservoirSample([rng], T, n::Int, method = algL; ordered = false)

Initializes a reservoir sample which can then be fitted with update!. The first signature represents a sample where only a single element is collected. Look at the Algorithms section for the supported methods.

StreamSampling.update! — Function

update!(rs::AbstractReservoirSample, el, [w])

Updates the reservoir sample by scanning the passed element. In the case of weighted sampling also the weight of the element needs to be passed to the function.

StreamSampling.value — Function

value(rs::AbstractReservoirSample)

Returns the elements collected in the sample at the current sampling stage.

StreamSampling.ordered_value — Function

ordered_value(rs::AbstractReservoirSample)

Returns the elements collected in the sample at the current sampling stage in the order they were collected. This applies only when ordered = true is passed in ReservoirSample.

StreamSampling.itsample — Function

itsample([rng], iter, method = algL)
itsample([rng], iter, weight, method = algAExpJ)

Return a random element of the iterator, optionally specifying a rng (which defaults to Random.default_rng()) and a weight function which accept each element as input and outputs the corresponding weight. If the iterator is empty, it returns nothing.

itsample([rng], iter, method = algL; ordered = false)
itsample([rng], iter, wv, method = algAExpJ; ordered = false)

Return a vector of n random elements of the iterator, optionally specifying a rng (which defaults to Random.default_rng()) and a method. ordered dictates whether an ordered sample (also called a sequential sample, i.e. a sample where items appear in the same order as in iter) must be collected.

If the iterator has less than n elements, in the case of sampling without replacement, it returns a vector of those elements.

Algorithms

StreamSampling.algL — Constant

Implements random sampling without replacement.

Adapted from algorithm L described in "Random sampling with a reservoir, J. S. Vitter, 1985".

StreamSampling.algR — Constant

Implements random sampling without replacement.

Adapted from algorithm R described in "Random sampling with a reservoir, J. S. Vitter, 1985".

StreamSampling.algRSWRSKIP — Constant

Implements random sampling with replacement.

Adapted fron algorithm RSWR_SKIP described in "Reservoir-based Random Sampling with Replacement from Data Stream, B. Park et al., 2008".

StreamSampling.algAExpJ — Constant

Implements weighted random sampling without replacement.

Adapted from algorithm A-ExpJ described in "Weighted random sampling with a reservoir, P. S. Efraimidis et al., 2006".

StreamSampling.algARes — Constant

Implements weighted random sampling without replacement.

Adapted from algorithm A-Res described in "Weighted random sampling with a reservoir, P. S. Efraimidis et al., 2006".

StreamSampling.algWRSWRSKIP — Constant

Implements weighted random sampling with replacement.

Adapted from algorithm WRSWR_SKIP described in "A Skip-based Algorithm for Weighted Reservoir Sampling with Replacement, A. Meligrana, 2024".